Add custom SQLAlchemy column options
This change allows the client to pass through custom SQLAlchemy column
options which will be added to the Column object on the SQLAlchemy
model. The specific need that motivated this change was to create
`created_at` and `updated_at` datetime fields. As such, this is what has
been used in the tests.